@@ -1,67 +1,79 @@
 {
-    "name": "Medicus",
-    "version": "1.0.0",
-    "description": "Medicus is a telemedicine application.",
-    "main": "app.js",
-    "scripts": {
-        "start": "node dist/app.js",
-        "devServer": "nodemon dist/app.js",
-        "buildServer": "rimraf dist/ && babel ./server --out-dir dist/ --copy-files",
-        "buildClient": "webpack",
-        "killNodeWin": "taskkill -F -IM node.exe",
-        "killNodeLinux": "killall .node.bin"
-    },
-    "license": "MIT",
-    "dependencies": {
-        "axios": "^0.20.0",
-        "bcryptjs": "^2.4.3",
-        "cors": "^2.8.5",
-        "dotenv": "^8.2.0",
-        "express": "^4.17.1",
-        "express-rate-limit": "^5.1.3",
-        "express-slow-down": "^1.3.1",
-        "helmet": "^4.1.1",
-        "jsonwebtoken": "^8.5.1",
-        "mongodb": "^3.6.4",
-        "multer": "^1.4.2",
-        "multer-gridfs-storage": "^4.2.0",
-        "react": "^16.12.0",
-        "react-dom": "^16.12.0",
-        "react-router-dom": "^5.1.2",
-        "react-transition-group": "^4.3.0",
-        "socket.io": "^3.1.2",
-        "socket.io-client": "^3.1.2"
-    },
-    "devDependencies": {
-        "@babel/cli": "^7.13.0",
-        "@babel/core": "^7.13.8",
-        "@babel/plugin-proposal-class-properties": "^7.13.0",
-        "@babel/plugin-proposal-decorators": "^7.13.5",
-        "@babel/plugin-proposal-do-expressions": "^7.12.13",
-        "@babel/plugin-proposal-export-default-from": "^7.12.13",
-        "@babel/plugin-proposal-export-namespace-from": "^7.12.13",
-        "@babel/plugin-proposal-function-sent": "^7.12.13",
-        "@babel/plugin-proposal-json-strings": "^7.13.8",
-        "@babel/plugin-proposal-logical-assignment-operators": "^7.13.8",
-        "@babel/plugin-proposal-nullish-coalescing-operator": "^7.13.8",
-        "@babel/plugin-proposal-numeric-separator": "^7.12.13",
-        "@babel/plugin-proposal-optional-chaining": "^7.13.8",
-        "@babel/plugin-proposal-pipeline-operator": "^7.12.13",
-        "@babel/plugin-proposal-throw-expressions": "^7.12.13",
-        "@babel/plugin-syntax-dynamic-import": "^7.8.3",
-        "@babel/plugin-syntax-import-meta": "^7.10.4",
-        "@babel/plugin-transform-runtime": "^7.13.8",
-        "@babel/preset-env": "^7.8.4",
-        "@babel/preset-react": "^7.8.3",
-        "@babel/register": "^7.13.8",
-        "@babel/runtime": "^7.13.9",
-        "babel-cli": "^6.26.0",
-        "babel-core": "^7.0.0-bridge.0",
-        "babel-loader": "^8.0.5",
-        "babel-preset-env": "^1.7.0",
-        "nodemon": "^2.0.7",
-        "rimraf": "^3.0.2",
-        "webpack": "^5.28.0",
-        "webpack-cli": "^4.6.0"
-    }
+  "name": "medicus",
+  "version": "0.1.0",
+  "private": true,
+  "dependencies": {
+    "@reduxjs/toolkit": "^1.5.1",
+    "@testing-library/jest-dom": "^4.2.4",
+    "@testing-library/react": "^9.5.0",
+    "@testing-library/user-event": "^7.2.1",
+    "bcryptjs": "^2.4.3",
+    "cors": "^2.8.5",
+    "dotenv": "^8.2.0",
+    "express": "^4.17.1",
+    "express-rate-limit": "^5.2.6",
+    "express-slow-down": "^1.4.0",
+    "helmet": "^4.5.0",
+    "jsonwebtoken": "^8.5.1",
+    "mongodb": "^3.6.6",
+    "multer": "^1.4.2",
+    "multer-gridfs-storage": "^4.2.0",
+    "react": "^17.0.2",
+    "react-dom": "^17.0.2",
+    "react-redux": "^7.2.3",
+    "react-responsive": "^8.2.0",
+    "react-router-dom": "^5.2.0",
+    "react-scripts": "4.0.3",
+    "socket.io": "^4.0.1",
+    "socket.io-client": "^4.0.1"
+  },
+  "scripts": {
+    "start": "node dist/app.js",
+    "test": "react-scripts test",
+    "eject": "react-scripts eject",
+    "buildClient": "react-scripts build",
+    "buildServer": "rimraf dist/ && babel ./server --out-dir dist/ --copy-files"
+  },
+  "eslintConfig": {
+    "extends": "react-app"
+  },
+  "browserslist": {
+    "production": [
+      ">0.2%",
+      "not dead",
+      "not op_mini all"
+    ],
+    "development": [
+      "last 1 chrome version",
+      "last 1 firefox version",
+      "last 1 safari version"
+    ]
+  },
+  "devDependencies": {
+    "@babel/cli": "^7.13.16",
+    "@babel/core": "^7.13.16",
+    "@babel/plugin-proposal-class-properties": "^7.13.0",
+    "@babel/plugin-proposal-decorators": "^7.13.15",
+    "@babel/plugin-proposal-do-expressions": "^7.12.13",
+    "@babel/plugin-proposal-export-default-from": "^7.12.13",
+    "@babel/plugin-proposal-export-namespace-from": "^7.12.13",
+    "@babel/plugin-proposal-function-sent": "^7.12.13",
+    "@babel/plugin-proposal-json-strings": "^7.13.8",
+    "@babel/plugin-proposal-logical-assignment-operators": "^7.13.8",
+    "@babel/plugin-proposal-nullish-coalescing-operator": "^7.13.8",
+    "@babel/plugin-proposal-numeric-separator": "^7.12.13",
+    "@babel/plugin-proposal-optional-chaining": "^7.13.12",
+    "@babel/plugin-proposal-pipeline-operator": "^7.12.13",
+    "@babel/plugin-proposal-throw-expressions": "^7.12.13",
+    "@babel/plugin-syntax-dynamic-import": "^7.8.3",
+    "@babel/plugin-syntax-import-meta": "^7.10.4",
+    "@babel/plugin-transform-runtime": "^7.13.15",
+    "@babel/preset-env": "^7.13.15",
+    "@babel/preset-react": "^7.13.13",
+    "@babel/register": "^7.13.16",
+    "@babel/runtime": "^7.13.17",
+    "node-sass": "^5.0.0",
+    "nodemon": "^2.0.7",
+    "rimraf": "^3.0.2"
+  }
 }
