@@ -1,56 +1,94 @@
 import React from 'react';
+import { useExtendClass } from "./hooks.js";
 
 
-export function Container(props) {
-  let className="container"
-  if (props.className) {
-    className = className.concat(" ", props.className);
-  }
-
-  return (
-    <div className={className}>{props.children}</div>
-  );
+export function Container(props) {  
+    return (
+        <div ref={props.ref} id={props.id} className={useExtendClass("container", props.className)}>
+            {props.children}
+        </div>
+    );
 }
+
 
 export function FluidContainer(props) {
-  let className="container-fluid"
-  if (props.className) {
-    className = className.concat(" ", props.className);
-  }
-
-  return (
-    <div className={className}>{props.children}</div>
-  );
+    return (
+        <div ref={props.ref} id={props.id} className={useExtendClass("container-fluid", props.className)}>
+            {props.children}
+        </div>
+    );
 }
+
 
 export function Row(props) {
-  let className="row"
-  if (props.className) {
-    className = className.concat(" ", props.className);
-  }
-
-  return (
-    <div className={className}>{props.children}</div>
-  );
+    return (
+        <div ref={props.ref} id={props.id} className={useExtendClass("row", props.className)}>
+            {props.children}
+        </div>
+    );
 }
 
-export function Col(props) {
-  let className="col"
-  if (props.className) {
-    className = className.concat(" ", props.className);
-  }
 
-  return (
-    <div className={className}>{props.children}</div>
-  );
+export function Col(props) {  
+    return (
+        <div ref={props.ref} id={props.id} className={useExtendClass("col", props.className)}>
+            {props.children}
+        </div>
+    );
 }
+
 
 export function RowDivider(props) {
-  return (
-    <Row className={props.className}>
-      <Col className="px-0">
-        <div className="border-top" />
-      </Col>
-    </Row>
-  );
+    return (
+        <Row {...props} >
+            <Col className="px-0">
+                <div className="border-top" />
+            </Col>
+        </Row>
+    );
+}
+
+
+export function Header(props) {
+    return (
+        <header ref={props.ref} id={props.id} className={useExtendClass("row", props.className)}>
+            {props.children}
+        </header>
+    );
+}
+
+
+export function Footer(props) {
+    return (
+        <footer ref={props.ref} id={props.id} className={useExtendClass("row", props.className)}>
+            {props.children}
+        </footer>
+    );
+}
+
+
+export function DescriptionList(props) {
+    return (
+        <dl ref={props.ref} id={props.id} className={useExtendClass("row", props.className)}>
+            {props.children}
+        </dl>
+    );
+}
+
+
+export function DescriptionTerm(props) {
+    return (
+        <dt ref={props.ref} id={props.id} className={useExtendClass("col", props.className)}>
+            {props.children}
+        </dt>
+    );
+}
+
+
+export function DescriptionDetails(props) {
+    return (
+        <dd ref={props.ref} id={props.id} className={useExtendClass("col", props.className)}>
+            {props.children}
+        </dd>
+    );
 }
