@@ -1,60 +1,62 @@
-import React from 'react';
-import { Switch, Route } from "react-router-dom";
-import { SessionContext } from '../../context/context.js';
+import React, { lazy, Suspense } from 'react';
+import { useSelector } from 'react-redux';
+import { Route, Switch } from 'react-router-dom';
+import { Loader } from '../../components/loaders';
 
-import AppointmentView from './appointmentView.js'
-import NewAppointmentView from './newAppointmentView.js';
-import { AllAppointmentsView, WaitingRoomView, PaymentsView } from './appointmentsView.js';
+const AppointmentView = lazy(() => import("./views/appointment"));
+const AllAppointmentsView = lazy(() => import("./views/allAppointments"));
+const WaitingRoomView = lazy(() => import('./views/waitingRoom'));
+const NewAppointmentView = lazy(() => import('./views/newAppointment'));
 
 
-export default class AppointmentApp extends React.Component {
-  constructor(props) {
-    super(props);
-  }
+export default function AppointmentApp(props) {
+    const session = useSelector(s => s.session);
 
-  render() {
     return (
-      <SessionContext.Consumer>
-        {sessionCtx => (
-          <Switch>
-            <Route
-              path={`${this.props.match.path}/:id`}
-              render={(props) => (
-                <AppointmentView {...props} session={sessionCtx.session} />
-              )}
-            />
-            <Route
-              path=""
-              render={(props) => {
-                if (props.location.search) {
-                  let searchParams = new URLSearchParams(this.props.location.search);
+        <Suspense fallback={<Loader isLoading={true} />}>
+            <Switch>
+                <Route
+                    path={`${props.match.path}/:id`}
+                    render={(props) => (
+                        <AppointmentView
+                            {...props} 
+                            id={props.match.params.id}
+                            session={session}
+                        />
+                    )}
+                />
+                <Route
+                    path=""
+                    render={(props) => {
+                        if (props.location.search) {
+                            let searchParams = new URLSearchParams(props.location.search);
 
-                  if (searchParams.has("view")) {
-                    const viewName = searchParams.get("view");
+                            if (searchParams.has("view")) {
+                                const viewName = searchParams.get("view");
 
-                    if (viewName === "new") {
-                      let physician = "";
-                      if (searchParams.has("physician")) {
-                        physician = searchParams.get("physician");
-                      }
+                                if (viewName === "new") {
+                                    let physician = "";
+                                    if (searchParams.has("physician")) {
+                                        physician = searchParams.get("physician");
+                                    }
 
-                      return (
-                        <NewAppointmentView {...props} session={sessionCtx.session} physician={physician} />
-                      );
-                    } else if (viewName === "waiting") {
-                      return (<WaitingRoomView {...props} session={sessionCtx.session} />);
-                    } else if (viewName === "payments") {
-                      return (<PaymentsView {...props} session={sessionCtx.session} />);
-                    }
-                  }
-                }
+                                    return (
+                                        <NewAppointmentView {...props} session={session} />
+                                    );
+                                } else if (viewName === "waiting") {
+                                    return (
+                                        <WaitingRoomView {...props} session={session} />
+                                    );
+                                }
+                            }
+                        }
 
-                return (<AllAppointmentsView {...props} session={sessionCtx.session} />);
-              }}
-            />
-          </Switch>
-        )}
-      </SessionContext.Consumer>
+                        return (
+                            <AllAppointmentsView {...props} session={session} />
+                        );
+                    }}
+                />
+            </Switch>
+        </Suspense>
     );
-  }
 }
