@@ -1,18 +1,17 @@
 import React from 'react';
 import { FluidContainer } from "../../components/layout.js";
-import HeaderBar from './header.js';
-import ContentBar from './content.js';
-import FooterBar from './footer.js';
+
+import HeaderBar from "./header.js";
+import ContentBar from "./content.js";
+import FooterBar from "./footer.js";
 
 
-export default class WelcomeApp extends React.Component {
-  render() {
+export default function WelcomeApp(props) {
     return (
-      <FluidContainer className="h-100 md-bg-0 d-flex flex-column">
-        <HeaderBar />
-        <ContentBar />
-        <FooterBar />
-      </FluidContainer>
+        <FluidContainer className="h-100 md-bg-0 d-flex flex-column">
+            <HeaderBar />
+            <ContentBar />
+            <FooterBar />
+        </FluidContainer>
     );
-  }
 }
