@@ -1,65 +1,82 @@
-import React from 'react';
+import React, { useState } from 'react';
+import { useDispatch } from 'react-redux';
 import { Form, FormRow, FormGroup, FormSubmit } from "../../components/form.js";
 
 
-export default class SignInForm extends React.Component {
-  constructor(props) {
-    super(props);
+export default function SignInForm(props) {
+    const dispatch = useDispatch();
 
-    this.state = {
-      errorMessage: "",
-      username: "",
-      password: ""
-    }
-
-    this.handleChange = this.handleChange.bind(this);
-    this.handleSubmit = this.handleSubmit.bind(this);
-  }
-
-  handleChange(e) {
-    event.preventDefault();
-    this.setState({
-      [e.target.name]: e.target.value
+    const [fields, setfields] = useState({
+        errorMessage: "",
+        username: "",
+        password: ""
     });
-  }
 
-  async handleSubmit(e) {
-    event.preventDefault();
-    try {
-      await this.props.signIn(this.state.username, this.state.password);
-    } catch (err) {
-      console.error(err);
-      this.setState({
-        errorMessage: err.message
-      });
+    function handleChange(e) {
+        setfields({
+            ...fields,
+            [e.target.name]: e.target.value
+        });
     }
-  }
 
-  render() {
-    return (
-      <Form handleSubmit={this.handleSubmit}>
-        <h5 className="font-weight-bold mb-3">Sign In.</h5>
-        {this.state.errorMessage &&
-          <FormRow className="justify-content-center">
-            <div className="alert alert-danger p-2" role="alert">{this.state.errorMessage}</div>
-          </FormRow>
+    async function handleSubmit(e) {
+        e.preventDefault();
+        try {
+            const response = await fetch(`/api/auth/signin`, {
+                method: 'POST',
+                credentials: 'same-origin',
+                headers: {
+                    'Content-Type': 'application/json'
+                },
+                body: JSON.stringify({
+                    username: fields.username,
+                    password: fields.password
+                })
+            });
+
+            let data = await response.json();
+            if (!response.ok) {
+                throw new Error(data.message);
+            }
+
+            dispatch({
+                type: "session/set",
+                payload: {
+                    ...data
+                }
+            });
+        } catch (err) {
+            console.error(`Failed to sign in. ${err}`);
+            setfields({
+                ...fields,
+                errorMessage: err.message
+            });
         }
-        <FormGroup>
-          <input id="signInForm1" type="text" name="username" className="form-control"
-            placeholder="Username" value={this.state.username} onChange={this.handleChange}
-              aria-label="Username" required>
-          </input>
-        </FormGroup>
-        <FormGroup>
-          <input id="signInForm2" type="password" name="password" className="form-control"
-            placeholder="Password" value={this.state.password} onChange={this.handleChange}
-              aria-label="Password" required>
-          </input>
-        </FormGroup>
-        <FormGroup className="mb-2">
-          <FormSubmit className="w-100">Sign In</FormSubmit>
-        </FormGroup>
-      </Form>
+    }
+
+    return (
+        <Form handleSubmit={handleSubmit}>
+            <h5 className="font-weight-bold mb-3">Sign In.</h5>
+            {fields.errorMessage &&
+                <FormRow className="justify-content-center">
+                    <div className="alert alert-danger p-2" role="alert">
+                        {fields.errorMessage}
+                    </div>
+                </FormRow>
+            }
+            <FormGroup>
+                <input id="signInForm1" type="text" name="username" className="form-control"
+                    placeholder="Username" value={fields.username} onChange={handleChange}
+                        aria-label="Username" required />
+            </FormGroup>
+            <FormGroup>
+                <input id="signInForm2" type="password" name="password" className="form-control"
+                    placeholder="Password" value={fields.password} onChange={handleChange}
+                        aria-label="Password" required />
+            </FormGroup>
+            <FormGroup className="mb-2">
+                <FormSubmit className="btn-primary w-100">Sign In</FormSubmit>
+            </FormGroup>
+        </Form>
     );
-  }
 }
